requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.57.0
grpcio==1.70.0
grpcio-tools==1.70.0
idna==3.10
iniconfig==2.1.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
protobuf==5.29.3
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
readerwriterlock==1.0.9
requests==2.32.3
six==1.17.0
typing_extensions==4.13.0
urllib3==2.3.0
